+++ src/map.rs
@@ -1,5 +1,5 @@
-mod map;
+mod map_systems;
 pub mod objectives;
 mod tile;
 
-pub use map::MapPlugin;
+pub use map_systems::MapPlugin;
